+++ modulo-03-1-springweb/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/service/PessoaService.java
@@ -1,66 +1,70 @@
 package br.com.dbc.vemser.pessoaapi.service;
 
+import br.com.dbc.vemser.pessoaapi.dto.PessoaCreateDTO;
+import br.com.dbc.vemser.pessoaapi.dto.PessoaDTO;
 import br.com.dbc.vemser.pessoaapi.entity.Pessoa;
 import br.com.dbc.vemser.pessoaapi.exceptions.RegraDeNegocioException;
 import br.com.dbc.vemser.pessoaapi.repository.PessoaRepository;
-import org.apache.commons.lang3.ObjectUtils;
-import org.apache.commons.lang3.StringUtils;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.Data;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@Data
 public class PessoaService {
 
+    // @Autowired
     private final PessoaRepository pessoaRepository;
+    private final ObjectMapper objectMapper;
 
-    public PessoaService(PessoaRepository pessoaRepository) {
-        this.pessoaRepository = pessoaRepository;
-    }
 
-    public Pessoa create(Pessoa pessoa) {
-//        if (StringUtils.isBlank(pessoa.getNome()))
-//        {
-//            System.out.println("Nome vazio. Erro!");
-//            return null;
-//        }
-//        if (ObjectUtils.isEmpty(pessoa.getDataNascimento())){
-//            System.out.println("Não é possível inserir uma pessoa sem data de nascimento. Erro!");
-//            return null;
-//        }
-//        if (pessoa.getCpf().length()<11 || StringUtils.isBlank(pessoa.getCpf()))
-//        {
-//            System.out.println("CPF inválido. Erro!");
-//            return null;
-//        }
-        return pessoaRepository.create(pessoa);
+    public PessoaDTO create(PessoaCreateDTO pessoa) throws RegraDeNegocioException {
+        Pessoa entity = objectMapper.convertValue(pessoa, Pessoa.class);
+        entity.setCpf(pessoa.getCpf());
+        entity.setNome(pessoa.getNome());
+        entity.setDataNascimento(pessoa.getDataNascimento());
+
+        Pessoa pessoa1 = pessoaRepository.create(entity);
+
+        PessoaDTO pessoaDTO = new PessoaDTO();
+        pessoaDTO.setIdPessoa(pessoa1.getIdPessoa());
+        pessoaDTO.setDataNascimento(pessoa1.getDataNascimento());
+        pessoaDTO.setCpf(pessoa1.getCpf());
+        pessoaDTO.setNome(pessoa1.getNome());
+
+        return pessoaDTO;
     }
 
     public List<Pessoa> list() {
         return pessoaRepository.list();
     }
 
-    public Pessoa update(Integer id,
-                         Pessoa pessoaAtualizar) throws Exception {
-        Pessoa pessoaRecuperada = getPessoa(id);
-
-        pessoaRecuperada.setCpf(pessoaAtualizar.getCpf());
-        pessoaRecuperada.setNome(pessoaAtualizar.getNome());
-        pessoaRecuperada.setDataNascimento(pessoaAtualizar.getDataNascimento());
-
-        return pessoaRecuperada;
+    public PessoaDTO update(Integer idPessoa,
+                         PessoaDTO pessoaAtualizar) throws RegraDeNegocioException {
+        pessoaAtualizar.setIdPessoa(idPessoa);
+        Pessoa entity = getPessoa(pessoaAtualizar.getIdPessoa());
+        if (entity != null) {
+            entity.setCpf(pessoaAtualizar.getCpf());
+            entity.setNome(pessoaAtualizar.getNome());
+            entity.setDataNascimento(pessoaAtualizar.getDataNascimento());
+        }
+        return pessoaAtualizar;
     }
 
-    public void delete(Integer id) throws Exception {
+    public void delete(Integer id) throws RegraDeNegocioException {
         Pessoa pessoaRecuperada = getPessoa(id);
-        pessoaRepository.delete(pessoaRecuperada);
+        if (pessoaRecuperada != null) {
+            pessoaRepository.delete(pessoaRecuperada);
+        }
     }
 
     public List<Pessoa> listByName(String nome) {
         return pessoaRepository.listByName(nome);
     }
 
-    public Pessoa getPessoa(Integer id) throws Exception {
+    public Pessoa getPessoa(Integer id) throws RegraDeNegocioException {
         Pessoa pessoaRecuperada = pessoaRepository.list().stream()
                 .filter(pessoa -> pessoa.getIdPessoa().equals(id))
                 .findFirst()
